third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferToGPUTexture-usage-flags.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./canvas-test.js"></script>
<script src="./webgpu-helpers.js"></script>
</head>

<h1>Canvas2D WebGPU Transfer: transferToGPUTexture-usage-flags</h1>

<p>transferToGPUTexture() creates a texture honoring the requested usage flags.
</p>

<script>

runCanvasTestsInWorker({dependencies: ['./webgpu-helpers.js']});

/**
 * transferToGPUTexture() should create a texture which honors the requested
 * usage flags.
 */
canvasPromiseTest(async (canvas) => {
  const adapter = await getWebGPUAdapter();
  // Skip this test on Mac Swiftshader due to "Invalid Texture" errors.
  if (isMacSwiftShader(adapter.info)) {
    return;
  }

  const device = await getWebGPUDevice(adapter);

  // Declare a helper function which tests individual usage flags.
  const testOneUsageFlag = async function(canvas, usageToEnable, usageToTest) {
    const ctx = canvas.getContext('2d');
    const tex = ctx.transferToGPUTexture({device: device,
                                          usage: usageToEnable});

    // Check that the GPUTexture's usage flags match our requested flags.
    assert_equals(tex.usage, usageToEnable);

    // Confirm that we _actually_ have the requested usage by performing an
    // action requiring that usage, and verifying no GPUValidationError occurs.
    device.pushErrorScope('validation');

    if (usageToTest & GPUTextureUsage.COPY_SRC) {
      copyOnePixelFromTextureAndSubmit(device, tex);
    }
    if (usageToTest & GPUTextureUsage.COPY_DST) {
      copyOnePixelToTextureAndSubmit(device, tex);
    }
    if (usageToTest & GPUTextureUsage.TEXTURE_BINDING) {
      createBindGroupUsingTexture(device, tex);
    }
    if (usageToTest & GPUTextureUsage.RENDER_ATTACHMENT) {
      clearTextureToColor(device, tex, { r: 1.0, g: 1.0, b: 1.0, a: 1.0 });
    }

    const errors = await device.popErrorScope();
    if (usageToTest == usageToEnable) {
      assert_equals(errors, null, 'enabled ' + usageToEnable +
                                  ', tested ' + usageToTest);
    } else {
      assert_not_equals(errors, null, 'enabled ' + usageToEnable +
                                      ', tested ' + usageToTest);
    }
    ctx.transferBackFromGPUTexture();
  };

  // Build up a chain of promises which test each TextureUsage flag.
  const supportedUsageFlags = [GPUTextureUsage.COPY_SRC,
                               GPUTextureUsage.COPY_DST,
                               GPUTextureUsage.TEXTURE_BINDING,
                               GPUTextureUsage.RENDER_ATTACHMENT];

  for (const usageToEnable of supportedUsageFlags) {
    for (const usageToTest of supportedUsageFlags) {
      await testOneUsageFlag(canvas, usageToEnable, usageToTest);
    }
  }
}, 'transferToGPUTexture() should create a texture which honors the ' +
   'requested usage flags.');

</script>
